+++ lib/chat_web/live/main_live/page/admin_panel.ex
@@ -28,6 +28,14 @@ defmodule ChatWeb.MainLive.Page.AdminPanel do
     |> assign_room_list()
   end
 
+  def int(socket) do
+    socket
+    |> assign(
+      need_login: true,
+      handshaked: false
+    )
+  end
+
   def request_wifi_settings(socket) do
     request_platform(:get_wifi_settings)