+++ CHANGELOG.md
@@ -18,6 +18,7 @@ straightforward as possible.
 - updating feed rendering w/ new mechanics
 
 ### Fixed
+- admin panel w/o login returns to login form (when page gets reconnected in admin panel)
 
 ## 2023-03-31_d8b27f7___2023-03-31_875d678